app/requirements.txt
Flask==3.0.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-exporter-jaeger==1.20.0
opentelemetry-instrumentation-flask==0.41b0
